adding sam signals

+++ sim_x10x20x10/run_skim_signal_bdt.sh
@@ -17,6 +17,11 @@ do
         POSITIONAL+=("$1")
         shift
         ;;
+        --sam)
+        SAM=true
+        POSITIONAL+=("$1")
+        shift
+        ;;
         *)    # unknown option
         POSITIONAL+=("$1") # save it in an array for later
         shift # past argument
@@ -34,6 +39,7 @@ module load cmssw
 cmsenv
 
 OUTPUT_DIR=$SKIM_SIG_BDT_OUTPUT_DIR
+INPUT_DIR=$SKIM_SIG_OUTPUT_DIR
 #OUTPUT_DIR="$OUTPUT_WD/signal/skim_signal_bdt_tighter"
 
 if [ -n "$SC" ]; then
@@ -42,25 +48,34 @@ if [ -n "$SC" ]; then
     OUTPUT_DIR=$SKIM_SIG_BDT_SC_OUTPUT_DIR
 fi
 
-#check output directory
-if [ ! -d "$OUTPUT_DIR" ]; then
-  mkdir $OUTPUT_DIR
-else
-   rm -rf $OUTPUT_DIR
-   mkdir $OUTPUT_DIR
+if [ -n "$SAM" ]; then
+    echo "GOT SAM"
+    echo "HERE: $@"
+    INPUT_DIR=$SKIM_SIG_SAM_OUTPUT_DIR
+    OUTPUT_DIR=$SKIM_SAM_SIG_BDT_OUTPUT_DIR
 fi
 
-#check output directory
-if [ ! -d "$OUTPUT_DIR/single" ]; then
-    mkdir "$OUTPUT_DIR/single"
-fi
+if [ -z "$SAM" ]; then
+    #check output directory
+    if [ ! -d "$OUTPUT_DIR" ]; then
+      mkdir $OUTPUT_DIR
+    else
+       rm -rf $OUTPUT_DIR
+       mkdir $OUTPUT_DIR
+    fi
 
-if [ ! -d "$OUTPUT_DIR/stdout" ]; then
-    mkdir "$OUTPUT_DIR/stdout"
-fi
+    #check output directory
+    if [ ! -d "$OUTPUT_DIR/single" ]; then
+        mkdir "$OUTPUT_DIR/single"
+    fi
 
-if [ ! -d "$OUTPUT_DIR/stderr" ]; then
-    mkdir "$OUTPUT_DIR/stderr"
+    if [ ! -d "$OUTPUT_DIR/stdout" ]; then
+        mkdir "$OUTPUT_DIR/stdout"
+    fi
+
+    if [ ! -d "$OUTPUT_DIR/stderr" ]; then
+        mkdir "$OUTPUT_DIR/stderr"
+    fi
 fi
 
 timestamp=$(date +%Y%m%d_%H%M%S%N)
@@ -75,52 +90,59 @@ notification = Never
 priority = 0
 EOM
 
-for sim in ${SKIM_SIG_OUTPUT_DIR}/sum/*; do
-    filename=`echo $(basename $sim .root) | awk -F"_" '{print $1"_"$2"_"$3}'`
-    echo $filename
-    tb=$filename
-    for group in "${!SIM_GROUP[@]}"; do
-        if [[ $group == "all" ]]; then
-            echo "Skipping ALL!!!!!"
-            continue
-        fi
-        echo checking group $group
-        value=${SIM_GROUP[$group]}
-        found=false
-        for pattern in $value; do
-            echo checking pattern $pattern
-            if [[ $filename == *"$pattern"* ]]; then
-                echo Found!
-                tb=$group
-                found=true
+if [ -z "$SAM" ]; then
+    for sim in ${INPUT_DIR}/sum/*; do
+        filename=`echo $(basename $sim .root) | awk -F"_" '{print $1"_"$2"_"$3}'`
+        echo $filename
+        tb=$filename
+        for group in "${!SIM_GROUP[@]}"; do
+            if [[ $group == "all" ]]; then
+                echo "Skipping ALL!!!!!"
+                continue
+            fi
+            echo checking group $group
+            value=${SIM_GROUP[$group]}
+            found=false
+            for pattern in $value; do
+                echo checking pattern $pattern
+                if [[ $filename == *"$pattern"* ]]; then
+                    echo Found!
+                    tb=$group
+                    found=true
+                    break
+                fi
+            done
+            if [[ "$found" = "true" ]]; then
                 break
             fi
         done
-        if [[ "$found" = "true" ]]; then
-            break
-        fi
-    done
-    echo "Will run:"
-    echo $CONDOR_WRAPPER $SCRIPTS_WD/skimmer_x1x2x1_univ_bdt_track_bdt.py -i $sim -o ${OUTPUT_DIR}/single/${filename}.root -tb $LEPTON_TRACK_SPLIT_DIR/cut_optimisation/tmva/$tb  -ub $OUTPUT_WD/cut_optimisation/tmva/total_bdt $@
-cat << EOM >> $output_file
+        echo "Will run:"
+        echo $CONDOR_WRAPPER $SCRIPTS_WD/skimmer_x1x2x1_univ_bdt_track_bdt.py -i $sim -o ${OUTPUT_DIR}/single/${filename}.root -tb $LEPTON_TRACK_SPLIT_DIR/cut_optimisation/tmva/$tb  -ub $OUTPUT_WD/cut_optimisation/tmva/total_bdt $@
+    cat << EOM >> $output_file
 arguments = $CONDOR_WRAPPER $SCRIPTS_WD/skimmer_x1x2x1_univ_bdt_track_bdt.py -i $sim -o ${OUTPUT_DIR}/single/${filename}.root -tb $LEPTON_TRACK_SPLIT_DIR/cut_optimisation/tmva/$tb  -ub $OUTPUT_WD/cut_optimisation/tmva/total_bdt $@
 error = ${OUTPUT_DIR}/stderr/${filename}.err
 output = ${OUTPUT_DIR}/stdout/${filename}.output
 Queue
 EOM
-done
+    done
+fi
 
 ## ALL GROUP
 
 echo -e "\n\nRUNNING ALL GROUP\n\n"
 
 OUTPUT_DIR=${SKIM_SIG_BDT_OUTPUT_DIR}_all
-#OUTPUT_DIR="$OUTPUT_WD/signal/skim_signal_bdt_tighter"
 
 if [ -n "$SC" ]; then
     echo "GOT SC"
     echo "HERE: $@"
     OUTPUT_DIR=${SKIM_SIG_BDT_SC_OUTPUT_DIR}_all
+fi
+
+if [ -n "$SAM" ]; then
+    echo "GOT SAM"
+    echo "HERE: $@"
+    OUTPUT_DIR=${SKIM_SAM_SIG_BDT_OUTPUT_DIR}_all
 fi
 
 #check output directory
@@ -144,8 +166,12 @@ if [ ! -d "$OUTPUT_DIR/stderr" ]; then
     mkdir "$OUTPUT_DIR/stderr"
 fi
 
-for sim in ${SKIM_SIG_OUTPUT_DIR}/sum/*; do
-    filename=`echo $(basename $sim .root) | awk -F"_" '{print $1"_"$2"_"$3}'`
+for sim in ${INPUT_DIR}/sum/*; do
+    if [ -n "$SAM" ]; then
+        filename=`echo $(basename $sim .root)`
+    else
+        filename=`echo $(basename $sim .root) | awk -F"_" '{print $1"_"$2"_"$3}'`
+    fi
     echo $filename
     tb=all
     echo "Will run:"
